Add a network suitable for testing the "Add" layer
Having added the add layer to the main repo, the next step is to
be able to run unit tests on it. This commit provides a basic network
that includes two add layers (adding 2 and 3 different layers together)
to be used in the test. The other layers in the network are just dense.